Add some helpful GitHub workflows (#2)
Added two new workflows to support PowerShell development and testing:

1. `PSScriptAnalyzer` to perform static code analysis
2. `PS Compatibility Matrix` to perform a quick compatibility check by executing the script using different builds of PowerShell

The latter took many attempts to get working due to nuances with GitHub workflows and PowerShell installation. Ultimately, it executes the script using Windows PowerShell 5.1 as well as the oldest and latest versions of PowerShell Core 6 and 7. In the future, I may try to get some PowerShell 2.0, 3.0, 4.0, and 5.0 checks in there (but I'm currently limited by GitHub's runners, and the script requires version 5.1 at this time).